--- a/posthog/admin/paginators/no_count_paginator.py
+++ b/posthog/admin/paginators/no_count_paginator.py
@@ -1,8 +1,7 @@
-from functools import cached_property
 from django.core.paginator import Paginator
 
 
 class NoCountPaginator(Paginator):
-    @cached_property
+    @property
     def count(self):
         return 999999